Split the ObjectManager into 3 classes (metadata, fragments/indexing, cache management/resolution)
